Factor knowledge of full extension into FullExtensionUtilities
