+++ src/GitHub.VisualStudio.16/GitHub.VisualStudio.16.csproj
@@ -67,10 +67,14 @@
     <Compile Include="Base\TeamExplorerBaseSection.cs" />
     <Compile Include="CompositionServices.cs" />
     <Compile Include="ExtensionServicesFactory.cs" />
+    <Compile Include="FullExtensionUtilities.cs" />
+    <Compile Include="Home\GitHubHomeContent.xaml.cs">
+      <DependentUpon>GitHubHomeContent.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="Home\HomeSection.cs" />
     <Compile Include="IExtensionServices.cs" />
     <Compile Include="Images.imagemanifest.cs" />
-    <Compile Include="InBoxExtensionServices.cs" />
-    <Compile Include="InstalledExtensionServices.cs" />
+    <Compile Include="InBoxGitHubContainerProvider.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Resources.Designer.cs">
       <AutoGen>True</AutoGen>
