+++ src/GitHub.VisualStudio.16/InBoxGitHubContainerProvider.cs
@@ -13,14 +13,6 @@ using System.Runtime.InteropServices;
 
 namespace GitHub.VisualStudio
 {
-    public class InBoxExtensionServices : IExtensionServices
-    {
-        public ICodeContainerProvider GetGitHubContainerProvider()
-        {
-            return new InBoxGitHubContainerProvider();
-        }
-    }
-
     [Guid(Guids.CodeContainerProviderId)]
     public class InBoxGitHubContainerProvider : ICodeContainerProvider
     {
@@ -37,12 +29,7 @@ namespace GitHub.VisualStudio
 
         async Task<CodeContainer> RunAcquisitionAsync(IProgress<ServiceProgressData> downloadProgress, CancellationToken cancellationToken, string url = null)
         {
-            await ThreadHelper.JoinableTaskFactory.SwitchToMainThreadAsync();
-
-            var componentModel = await ServiceProvider.GetGlobalServiceAsync<SComponentModel, IComponentModel>();
-            Assumes.Present(componentModel);
-
-            var result = await ShowCloneDialogAsync(componentModel, downloadProgress, cancellationToken, url);
+            var result = await ShowCloneDialogAsync(downloadProgress, cancellationToken, url);
             if (result == null)
             {
                 return null;
@@ -69,9 +56,13 @@ namespace GitHub.VisualStudio
             return codeContainer;
         }
 
-        static async Task<CloneDialogResult> ShowCloneDialogAsync(IComponentModel componentModel,
-            IProgress<ServiceProgressData> downloadProgress, CancellationToken cancellationToken, string url = null)
+        async Task<CloneDialogResult> ShowCloneDialogAsync(IProgress<ServiceProgressData> downloadProgress,
+            CancellationToken cancellationToken, string url = null)
         {
+            await ThreadHelper.JoinableTaskFactory.SwitchToMainThreadAsync();
+
+            var componentModel = await ServiceProvider.GetGlobalServiceAsync<SComponentModel, IComponentModel>();
+            Assumes.Present(componentModel);
             var compositionServices = componentModel.DefaultExportProvider.GetExportedValue<CompositionServices>();
             var exportProvider = compositionServices.GetExportProvider();
