+++ src/GitHub.VisualStudio.16/Sync/PublishSection.cs
@@ -10,6 +10,7 @@ using GitHub.Primitives;
 using GitHub.ViewModels.TeamExplorer;
 using GitHub.VisualStudio;
 using Microsoft.TeamFoundation.Controls;
+using Microsoft.VisualStudio.Shell;
 using ReactiveUI;
 
 
@@ -27,7 +28,6 @@ namespace Microsoft.TeamExplorerSample.Sync
         public const int Priority = 4;
 
         readonly Guid PushToRemoteSectionId = new Guid("99ADF41C-0022-4C03-B3C2-05047A3F6C2C");
-        readonly Guid GitHubPublishSectionId = new Guid("92655B52-360D-4BF5-95C5-D9E9E596AC76");
 
         [ImportingConstructor]
         public PublishSection(CompositionServices compositionServices)
@@ -58,7 +58,16 @@ namespace Microsoft.TeamExplorerSample.Sync
         /// </summary>
         public override void Initialize(object sender, SectionInitializeEventArgs e)
         {
+            ThreadHelper.ThrowIfNotOnUIThread();
+
             base.Initialize(sender, e);
+
+            // Remain hidden when full GitHub extension is installed
+            if (FullExtensionUtilities.IsInstalled(ServiceProvider))
+            {
+                IsVisible = false;
+                return;
+            }
 
             RefreshVisibility();
 
@@ -67,11 +76,6 @@ namespace Microsoft.TeamExplorerSample.Sync
                 if (page.GetSection(PushToRemoteSectionId) is ITeamExplorerSection pushToRemoteSection)
                 {
                     pushToRemoteSection.PropertyChanged += Section_PropertyChanged;
-                }
-
-                if (page.GetSection(GitHubPublishSectionId) is ITeamExplorerSection gitHubPublishSection)
-                {
-                    gitHubPublishSection.PropertyChanged += Section_PropertyChanged;
                 }
             }
 
@@ -123,27 +127,27 @@ namespace Microsoft.TeamExplorerSample.Sync
 
         void RefreshVisibility()
         {
-            bool IsSectionVisible(ITeamExplorerPage teamExplorerPage, Guid sectionId)
-            {
-                if (teamExplorerPage.GetSection(sectionId) is ITeamExplorerSection pushToRemoteSection)
-                {
-                    return pushToRemoteSection.SectionContent != null && pushToRemoteSection.IsVisible;
-                }
-
-                return false;
-            }
-
             var visible = false;
 
             if (ServiceProvider.GetService(typeof(ITeamExplorerPage)) is ITeamExplorerPage page)
             {
-                visible = IsSectionVisible(page, PushToRemoteSectionId) && !IsSectionVisible(page, GitHubPublishSectionId);
+                visible = IsSectionVisible(page, PushToRemoteSectionId);
             }
 
             if (IsVisible != visible)
             {
                 IsVisible = visible;
             }
+        }
+
+        bool IsSectionVisible(ITeamExplorerPage teamExplorerPage, Guid sectionId)
+        {
+            if (teamExplorerPage.GetSection(sectionId) is ITeamExplorerSection pushToRemoteSection)
+            {
+                return pushToRemoteSection.SectionContent != null && pushToRemoteSection.IsVisible;
+            }
+
+            return false;
         }
     }
 }
